requirements.txt
numpy==1.18.4
pandas==1.0.3
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
snips-nlu-parsers==0.4.3
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
tokenizers==0.7.0
tqdm==4.46.0
transformers==2.9.0
